+++ Amplitude/Amplitude.m
@@ -288,7 +288,19 @@ static NSString *const SEQUENCE_NUMBER = @"sequence_number";
     // migrate events
     NSArray *events = [eventsData objectForKey:EVENTS];
     for (id event in events) {
-        NSData *jsonData = [NSJSONSerialization dataWithJSONObject:event options:0 error:NULL];
+        NSError *error = nil;
+        NSData *jsonData = nil;
+        @try {
+            jsonData = [NSJSONSerialization dataWithJSONObject:[self makeJSONSerializable:event] options:0 error:&error];
+        }
+        @catch (NSException *exception) {
+            NSLog(@"ERROR: NSJSONSerialization error: %@", exception.reason);
+            continue;
+        }
+        if (error != nil) {
+            NSLog(@"ERROR: NSJSONSerialization error: %@", error);
+            continue;
+        }
         NSString *jsonString = [[NSString alloc] initWithData:jsonData encoding:NSUTF8StringEncoding];
         success &= [dbHelper addEvent:jsonString];
         SAFE_ARC_RELEASE(jsonString);